Update PlanetClass.py
Changed the way trails work. Now, instead of just not clearing the screen every time, the program stores data of the locations of the planets for the previous 1000 frames, and then displays a line connecting those points. This is useful because it is now possible to draw new planets while keeping trails active, and it also looks more continuous/clean.